pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "joblib>=1.5.1",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "xgboost>=3.0.2",
]
